% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemo_oncdrs_feature_gen.R
\name{chemo_oncdrs_feature_gen}
\alias{chemo_oncdrs_feature_gen}
\title{Generate DFCI chemotherapy-related features (chemo_oncdrs data).}
\usage{
chemo_oncdrs_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\description{
\
}

