% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_subset_data_H.R
\name{combine_subset_data}
\alias{combine_subset_data}
\title{Combine, subset datafiles for a given feature set}
\usage{
combine_subset_data(file_source_list, feature_set, cohort_empi)
}
\arguments{
\item{file_source_list}{list of paths to .Rds files which are expected to be "master datasets" for multiple patient populations (eg. list of dia files for icmp, bwh_ed_100k cohorts) (list of character)}

\item{feature_set}{abbreviation of feature set these files are relevant to (eg. "dia", "prc", "enc") (character)}

\item{cohort_empi}{vector of all unique empis in cohort observation set (character vector)}
}
\description{
For a a given feature set and list of data files (of master data from several cohorts), create a combined master subset containing only observations relevant to the provided cohort empi list.
}
