% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_feature_gen_F.R
\name{dem_feature_gen}
\alias{dem_feature_gen}
\title{Generate demographics-related features (dem data).}
\usage{
dem_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\arguments{
\item{cohort}{cohort dt}

\item{cohort_key_var_merge}{list of key identifiers on which to merge (see: feature_initialisation)}

\item{cohort_key_var}{list of key identifiers on which to merge (see: feature_initialisation)}
}
\description{
\
}
