% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia_feature_code.R
\name{dia_feature_gen}
\alias{dia_feature_gen}
\title{Generate diagnosis-related features (dia data).}
\usage{
dia_feature_gen(dia_file_mod = dia_file_mod, leak_dia_day = leak_dia_day,
  combine = FALSE, dia_file_mod_ext = NA, file_date_var = "dia_date")
}
\description{
\
}
