% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia_oncdrs_rpdr_feature_gen.R
\name{dia_oncdrs_rpdr_feature_gen}
\alias{dia_oncdrs_rpdr_feature_gen}
\title{Generate DFCI & RPDR (combined) diagnosis-related features (dia_oncdrs & dia data).}
\usage{
dia_oncdrs_rpdr_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\arguments{
\item{cohort_key_var_merge}{}
}
\description{
\
}
