% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_feature_creation_H.R
\name{diff_feature_creation}
\alias{diff_feature_creation}
\title{Collapse features into a category-level tabulation.}
\usage{
diff_feature_creation(dt, dt_feature, feature_list, feature_name,
  feature_name_ext, name_ext_tmp)
}
\arguments{
\item{dt}{}

\item{dt_feature}{}

\item{feature_list}{}

\item{feature_name}{}
}
\description{
\
}
