% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_feature_gen.R
\name{ed_feature_gen}
\alias{ed_feature_gen}
\title{Generate ED-related features (ed_admin data).}
\usage{
ed_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  order_file_date_var = "order_enter_date", enc_file_date_var = "adm_date")
}
\description{
\
}
