% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc_feature_gen_F.R
\name{enc_feature_gen}
\alias{enc_feature_gen}
\title{Generate enc-related features (enc data).}
\usage{
enc_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  file_date_var = "adm_date", leak_enc_day_arg = leak_enc_day,
  enc_file_mod_arg = enc_file_mod)
}
\description{
\
}
