% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc_oncdrs_feature_gen_F.R
\name{enc_oncdrs_feature_gen}
\alias{enc_oncdrs_feature_gen}
\title{Generate DFCI enc-related features (enc_oncdrs data).}
\usage{
enc_oncdrs_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  file_date_var = "adm_date", leak_enc_oncdrs_day_arg = leak_enc_day,
  enc_oncdrs_file_mod_arg = enc_oncdrs_file_mod)
}
\description{
\
}
