% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_check.R
\name{feature_check}
\alias{feature_check}
\title{Verify the correct compilation of a feature set.}
\usage{
feature_check(feature_dt_name, cohort_dt = cohort,
  cohort_key_var_list = cohort_key_var,
  extra_var_list = names(cohort_extra_col), save_name = TRUE)
}
\arguments{
\item{feature_dt_name}{}

\item{cohort_dt}{}

\item{cohort_key_var_list}{}

\item{extra_var_list}{}

\item{save_name}{}
}
\description{
\
}
