% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_coll.R
\name{feature_coll}
\alias{feature_coll}
\title{Collapse features into a category-level tabulation.}
\usage{
feature_coll(dt, cohort_key_var_list = cohort_key_var,
  cohort_extra_var_list = names(cohort_extra_col))
}
\arguments{
\item{dt}{}

\item{cohort_key_var_list}{}

\item{cohort_extra_var_list}{}
}
\description{
\
}

