% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_compilation_M.R
\name{feature_compilation}
\alias{feature_compilation}
\title{Compile features (Stage II).}
\usage{
feature_compilation(cohort_path, control_path, data_def_path, feature_path,
  feature_set_id, feature_set_prefix)
}
\arguments{
\item{cohort_path}{}

\item{control_path}{}

\item{data_def_path}{}

\item{feature_path}{}

\item{feature_set_id}{}

\item{feature_set_prefix}{}
}
\description{
\
}
