% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_construction_M.R
\name{feature_construction}
\alias{feature_construction}
\title{Assemble features (Stage I).}
\usage{
feature_construction(cohort_path, control_path, data_path, feature_path,
  feature_set_id, feature_set_prefix)
}
\arguments{
\item{cohort_path}{}

\item{control_path}{}

\item{data_path}{}

\item{feature_path}{}

\item{feature_set_id}{}

\item{feature_set_prefix}{}
}
\description{
\
}
