% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_dataset_check_H.R
\name{feature_dataset_check}
\alias{feature_dataset_check}
\title{Verify the correct construction of a feature category dataset.}
\usage{
feature_dataset_check(feature_dt, feature_dt_name, cohort_dt,
  cohort_key_var_list, extra_var_list, save_name = TRUE)
}
\arguments{
\item{feature_dt_name}{}

\item{cohort_dt}{}

\item{cohort_key_var_list}{}

\item{extra_var_list}{}

\item{save_name}{}
}
\description{
\
}
