% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_modification_M.R
\name{feature_modification}
\alias{feature_modification}
\title{Modify features constructed in stage 1 [this is stage 2].}
\usage{
feature_modification(cohort_path, control_path, data_def_path, feature_path,
  feature_set_id, feature_set_prefix)
}
\arguments{
\item{cohort_path}{path to .Rds cohort file which must have the following column names - "empi", "outcome", "t0_date", "outcome_id"; can also contain a "test_set" column; other columns are non-essential but not a problem (character)}

\item{control_path}{path to control.R file which contains the user defined parameters for feature construction (character)}

\item{data_def_path}{path to data_def_path.R file which contains lists of relevant file paths where master data sets that are to be used in the FC process live (character)}

\item{feature_path}{path to feature_selection.csv file which provides options to select variables for FC process (character)}

\item{feature_set_id}{suffix indicating an id number or index number for this cohort's feature set (if multiple attempts were made with the same cohort, for example) (character)}

\item{feature_set_prefix}{prefix identifying the cohort / feature set that is used in the output file names for all datasets / output created as part of the FC process (character)}
}
\description{
Taking the feature sets constructed in stage 1 and saving each individual set in a .Rds file after modifying them by (a) selection - subsetting to only variables specified by the user in the feature_selection.csv (b) thresholding - dropping feature sets that do not meet the thresholds for missingness / zeroness as specified in the control.R file (c) missing imputation - reversing the automatic imputation (set missing instead of 0) from stage 1 for variables as specified in control.R (d) median imputation - imputing missing observations to the median value for the the features specified in control.R
}
