% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_overview_H.R
\name{feature_overview}
\alias{feature_overview}
\title{Initialise the feature construction machinery.}
\usage{
feature_overview(pred_set_final, pred_set, deselect_col, na_col, zero_col,
  date_col_table)
}
\arguments{
\item{pred_set_final}{Final feature set generated for which to generate an extensive set of summary stats}

\item{pred_set}{"Penultimate" feature set from which the final feature set is generated after some variable name changes and column subsetting}

\item{deselect_col}{Columns chosen to be deselected from the feature set}

\item{na_col}{Columns/features dropped based on the na threshold (quant missing threshold)}

\item{zero_col}{Columns/features dropped based on the zero threshold (indic_missing_threshold)}

\item{date_col_table}{data.table with earliest/latest observation dates for each of the feature sets}
}
\description{
\
}
