% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_sourcing_M.R
\name{feature_sourcing}
\alias{feature_sourcing}
\title{Source features (Stage 0).}
\usage{
feature_sourcing(cohort_path, control_path, data_def_path, feature_path,
  feature_set_id, feature_set_prefix)
}
\arguments{
\item{cohort_path}{path to .Rds cohort file which must have the following column names - "empi", "outcome", "t0_date", "outcome_id"; can also contain a "test_set" column; other columns are non-essential but not a problem (character)}

\item{control_path}{path to control.R file which contains the user defined parameters for feature construction (character)}

\item{data_def_path}{path to data_def_path.R file which contains lists of relevant file paths where master data sets that are to be used in the FC process live (character)}

\item{feature_path}{path to feature_selection.csv file which provides options to select variables for FC process (character)}

\item{feature_set_id}{suffix indicating an id number or index number for this cohort's feature set (if multiple attempts were made with the same cohort, for example) (character)}

\item{feature_set_prefix}{prefix identifying the cohort / feature set that is used in the output file names for all datasets / output created as part of the FC process (character)}
}
\description{
Create master subsets for each file type from which features are to be assembled. These master subsets are then used in Stage 1 to create features. The idea is to load in all master relevant master cohorts for each file type, subset them to include only records relevant to the cohort at hand, combine these subsetted datasets for each file type, then saving just one subsetted and merged master dataset for each file type, which acts as the source for all feature construction for that feature type. Doing this once allows rapid iteration in stages 1 and 2, since stage 0 is unlikely be repeated more than once for the feature construction process for a given cohort.
}
