% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_structure_vis.R
\name{feature_structure_vis}
\alias{feature_structure_vis}
\title{Generate a child-node file used to visualise the features as a tree diagram.}
\usage{
feature_structure_vis(var_name_raw_list = var_name_vis)
}
\arguments{
\item{var_name_raw_list}{}
}
\description{
\
}
