% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_type_format_H.R
\name{feature_type_format}
\alias{feature_type_format}
\title{Format features based on the variable type.}
\usage{
feature_type_format(dt, num_feature_pattern = NA, int_feature_pattern = NA,
  factor_feature_pattern = NA, day_to_last)
}
\arguments{
\item{dt}{}

\item{day_to_last}{}
}
\description{
\
}
