% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_var_format.R
\name{feature_var_format}
\alias{feature_var_format}
\title{Format features.}
\usage{
feature_var_format(dt_list, function_ext_1 = "_length_",
  function_ext_2 = "_function_")
}
\arguments{
\item{dt_list}{}
}
\description{
\
}

