% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_var_format_2.R
\name{feature_var_format_2}
\alias{feature_var_format_2}
\title{Format features.}
\usage{
feature_var_format_2(dt, timeframe_split_var = time_split)
}
\arguments{
\item{dt}{}

\item{timeframe_split_var}{}
}
\description{
\
}
