% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_var_format_day_to_last_H.R
\name{feature_var_format_day_to_last}
\alias{feature_var_format_day_to_last}
\title{Format features.}
\usage{
feature_var_format_day_to_last(dt,
  timeframe_split_var = timeframe_day_to_last)
}
\arguments{
\item{dt}{}

\item{timeframe_split_var}{}
}
\description{
\
}
