% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_data_cdc_data.R
\docType{data}
\name{flu_data_cdc}
\alias{flu_data_cdc}
\title{CDC flu data.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 19185 rows and 3 columns.}
\usage{
data(flu_data_cdc)
}
\description{
\
}
\keyword{datasets}
