% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete_hist.R
\name{incomplete_hist}
\alias{incomplete_hist}
\title{Visualise missingness across features.}
\usage{
incomplete_hist(dt, num_factor_var_list, indic_var_list,
  output_folder_path = output_folder, cohort_name_path = cohort_name,
  version_path)
}
\arguments{
\item{dt}{}

\item{num_factor_var_list}{}

\item{indic_var_list}{}

\item{output_folder_path}{}

\item{cohort_name_path}{}

\item{version_path}{}
}
\description{
\
}
