% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_oncdrs_rpdr_feature_gen.R
\name{lab_oncdrs_rpdr_feature_gen}
\alias{lab_oncdrs_rpdr_feature_gen}
\title{Generate DFCI & RPDR (combined) lab-related features (lab_oncdrs & lab data).}
\usage{
lab_oncdrs_rpdr_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\description{
\
}

