% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_chemo_oncdrs_rpdr_feature_gen.R
\name{med_chemo_oncdrs_rpdr_feature_gen}
\alias{med_chemo_oncdrs_rpdr_feature_gen}
\title{Generate DFCI & RPDR (combined) medication and chemotherapy-related 
features (med_oncdrs & med & chemo_oncdrs data).}
\usage{
med_chemo_oncdrs_rpdr_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\description{
\
}

