% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_feature_gen.R
\name{med_feature_gen}
\alias{med_feature_gen}
\title{Generate medication-related features (med data).}
\usage{
med_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  med_file_mod_arg = med_file_mod, leak_med_day_arg = leak_med_day,
  combine = FALSE, med_file_mod_ext = NA, med_file_mod_ext_ext = NA,
  file_date_var = "med_date")
}
\description{
\
}
