% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_oncdrs_feature_gen.R
\name{med_oncdrs_feature_gen}
\alias{med_oncdrs_feature_gen}
\title{Generate DFCI medication-related features (med_oncdrs data).}
\usage{
med_oncdrs_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\description{
\
}
