% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_feature_gen_F.R
\name{mic_feature_gen}
\alias{mic_feature_gen}
\title{Generate microbiology-related features (mic data).}
\usage{
mic_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  file_date_var = "mic_date", mic_file_mod_arg = mic_file_mod,
  leak_mic_day_arg = leak_mic_day)
}
\description{
\
}
