% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_check.R
\name{obs_check}
\alias{obs_check}
\title{Verify the correct compilation of a feature set.}
\usage{
obs_check(dt, cohort_key_var_list = cohort_key_var,
  cohort_extra_var_list = names(cohort_extra_col))
}
\arguments{
\item{dt}{}

\item{cohort_key_var_list}{}

\item{cohort_extra_var_list}{}
}
\description{
\
}
