% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_dependencies_H.R
\docType{data}
\name{return_mult}
\alias{return_mult}
\title{Store multiple arguments returned from function as separate objects - structure.}
\format{An object of class \code{result} of length 1.}
\usage{
return_mult
}
\description{
Store multiple arguments returned from function as separate objects - structure.
}
\keyword{datasets}
