% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_dependencies_H.R
\name{[<-.result}
\alias{[<-.result}
\title{Store multiple arguments returned from function as separate objects - function.}
\usage{
\method{[}{result}(x, ...) <- value
}
\description{
Store multiple arguments returned from function as separate objects - function.
}
