% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeframe_combine_H.R
\name{timeframe_combine}
\alias{timeframe_combine}
\title{Combine a dataframe split into different timeframes.}
\usage{
timeframe_combine(DT_list, DT_list_name, var_list = cohort_key_var)
}
\arguments{
\item{DT_list}{}
}
\description{
\
}
