% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_list_default_data.R
\docType{data}
\name{variable_list_default}
\alias{variable_list_default}
\title{Feature groups - used to select/subset features.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 178 rows and 6 columns.}
\usage{
data(variable_list_default)
}
\description{
\
}
\keyword{datasets}

