% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_class_data.R
\docType{data}
\name{zip_class}
\alias{zip_class}
\title{US zip code data.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 44336 rows and 3 columns.}
\usage{
data(zip_class)
}
\description{
\
}
\keyword{datasets}

