% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_feature_feature_gen.R
\name{dem_feature_gen}
\alias{dem_feature_gen}
\title{Generate demographics-related features (dem data).}
\usage{
dem_feature_gen(cohort, cohort_key_var_merge, cohort_key_var)
}
\description{
\
}

