% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc_feature_gen.R
\name{enc_feature_gen}
\alias{enc_feature_gen}
\title{Generate enc-related features (enc data).}
\usage{
enc_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  file_date_var = "adm_date")
}
\description{
\
}
