% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_compilation.R
\name{feature_compilation}
\alias{feature_compilation}
\title{Compile features (Stage II).}
\usage{
feature_compilation(control_file_path)
}
\arguments{
\item{control_file_path}{Path to control file (see template incl. in package)}
}
\description{
\
}
