% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_construction.R
\name{feature_construction}
\alias{feature_construction}
\title{Generate features (Stage I).}
\usage{
feature_construction(control_file_path)
}
\arguments{
\item{control_file_path}{Path to control file (see template incl. in package)}
}
\description{
\
}

