% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_feature_gen.R
\name{lab_feature_gen}
\alias{lab_feature_gen}
\title{Generate lab-related features (lab data).}
\usage{
lab_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  lab_file_mod_arg = lab_file_mod, leak_lab_day_arg = leak_lab_day,
  combine = FALSE, lab_file_mod_ext = NA, file_date_var = "lab_date")
}
\description{
\
}
