% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvs_feature_gen.R
\name{lvs_feature_gen}
\alias{lvs_feature_gen}
\title{Generate vital sign-related features (lvs data).}
\usage{
lvs_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  lvs_file_mod_arg = lvs_file_mod, leak_lvs_day_arg = leak_lvs_day,
  combine = FALSE, lvs_file_mod_ext = NA, lvs_file_mod_ext_ext = NA,
  load = TRUE, file_date_var = "lvs_date")
}
\description{
\
}

