% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_feature_code.R
\name{med_feature_gen}
\alias{med_feature_gen}
\title{Generate medication-related features (med data).}
\usage{
med_feature_gen(med_file_mod = med_file_mod, leak_med_day = leak_med_day,
  combine = FALSE, med_file_mod_ext = NA, med_file_mod_ext_ext = NA,
  file_date_var = "med_date")
}
\description{
\
}
