% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_feature_gen.R
\name{prc_feature_gen}
\alias{prc_feature_gen}
\title{Generate procedure-related features (prc data).}
\usage{
prc_feature_gen(cohort, cohort_key_var_merge, cohort_key_var,
  file_date_var = "prc_date")
}
\arguments{
\item{file_date_var}{}
}
\description{
/
}
