% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeframe_combine.R
\name{timeframe_combine}
\alias{timeframe_combine}
\title{Combine data split into different timeframes.}
\usage{
timeframe_combine(DT_list, var_list = cohort_key_var)
}
\arguments{
\item{DT_list}{}
}
\description{
\
}
